// packages/app1/src/enums/HttpEnum.ts
export enum ContentType {
	JSON = 'application/json;charset=UTF-8',
	FORM_URLENCODED = 'application/x-www-form-urlencoded;charset=UTF-8;',
	FORM_DATA = 'multipart/form-data;charset=UTF-8',
}

export enum RequestEnum {
	GET = 'GET',
	POST = 'POST',
	DELETE = 'DELETE',
	PUT = 'PUT',
	UPDATE = 'UPDATE',
}